Move OidcLogoutTokenAuthenticationConverter
